LayoutTests/platform/ios/imported/w3c/web-platform-tests/svg/coordinate-systems/svgtransformlist-replaceitem-expected.txt

123
123

FAIL Dynamic update of transform; replaceItem() assert_equals: expected 185.703125 but got 185.5
